Temp fix for Z-up to Y-up conversion
prism only outputs the transform matrix now. Need to revisit the conversion some other time.
An additional node transform is concatenated for mesh/skeletal graph nodes. This takes care
of the Z-up to Y-up conversion (read from the model) for now.

Adjustments had to be made for the physics to create the collision with the transformed
vertices.